break up handler logic and instrumentation configuration into separate files
